\name{idmsurv-package}
\alias{idmsurv-package}
\alias{idmsurv}
\docType{package}
\title{idmsurv: \packageTitle{idmsurv}
%A package for inference and prediction in an Illness-Death Model
}
\description{
Newly developed methods for the estimation of
of probabilities in an illness-death model.
FALTA
}


\author{
Luis Meira-Machado and Marta Sestelo.

Maintainer: Marta Sestelo, sestelo@uvigo.es
}

\references{
FALTA

}


