\name{survidm-package}
\alias{survidm-package}
\alias{survidm}
\docType{package}
\title{survidm: \packageTitle{survidm}
%A package for inference and prediction in an Illness-Death Model
}
\description{
Newly developed methods for the estimation of several probabilities in an illness-death model. The package can be used to obtain nonparametric and semiparametric estimates for: transition probabilities, occupation probabilities, cumulative incidence function and the sojourn time distributions. Additionally, it is possible to fit proportional hazards regression models in each transition of the
Illness-Death Model. Several auxiliary functions are also provided which can be used for marginal estimation of the survival functions.
}


\author{
Luis Meira-Machado and Marta Sestelo.

Maintainer: Marta Sestelo, sestelo@uvigo.es
}

\references{
Aalen O. O., Johansen S. (1978) An Empirical Transition Matrix for Nonhomogeneous Markov
Chains Based on Censored Observations. Scandinavian Journal of Statistics, 5(3), 141--150.

Cao R., Lopez-de Ullibarri I., Janssen P., Veraverbeke N. (2005) Presmoothed kaplan-meier and
nelsonaalen estimators. Journal of Nonparametric Statistics, 17, 31--56.

de Una-Alvarez J. and Meira-Machado L. (2015). Nonparametric estimation of transition probabilities in a non-Markov illness-death model: a comparative study. Biometrics, 71, 364--375.

Geskus R.B. (2011). Cause-specific cumulative incidence estimation and the fine and gray model under both left truncation and right censoring. Biometrics, 67, 39--49.

Meira-Machado L. F., de Una-Alvarez J. and Cadarso-Suarez C. (2006). Nonparametric estimation
of transition probabilities in a non-Markov illness-death model. Lifetime Data Anal 12(3), 325--344.

Satten G.A. and Datta S. (2002) Marginal estimation for multi-stage models: waiting time distributions and competing risks analyses. Statistics in Medicine, 21, 3--19.

}


